+++ ray-operator/config/manager/kustomization.yaml
@@ -1,7 +1,3 @@
-images:
-- name: controller
-  newName: seedjeffwan/ray-operator
-  newTag: v0.1.0
 resources:
 - manager.yaml
 apiVersion: kustomize.config.k8s.io/v1beta1
